Replace CDK LocalStack test base
