+++ java/cdk/src/test/java/sleeper/cdk/util/NewInstanceValidatorIT.java
@@ -19,8 +19,8 @@ package sleeper.cdk.util;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.io.TempDir;
 
-import sleeper.cdk.testutils.LocalStackTestBase;
 import sleeper.core.properties.instance.InstanceProperties;
+import sleeper.localstack.test.LocalStackTestBase;
 import sleeper.statestore.dynamodb.DynamoDBStateStore;
 import sleeper.statestore.dynamodb.DynamoDBStateStoreCreator;
 import sleeper.statestore.s3.S3StateStore;
@@ -82,7 +82,7 @@ class NewInstanceValidatorIT extends LocalStackTestBase {
     @Test
     void shouldThrowAnErrorWhenDynamoStateStoreExists() throws IOException {
         // Given
-        new DynamoDBStateStoreCreator(instanceProperties, dynamoClientV1).create();
+        new DynamoDBStateStoreCreator(instanceProperties, dynamoDBClient).create();
         setupTablesPropertiesFile(temporaryFolder, "example-table", DynamoDBStateStore.class.getName());
 
         // When
@@ -94,7 +94,7 @@ class NewInstanceValidatorIT extends LocalStackTestBase {
     @Test
     void shouldThrowAnErrorWhenS3StateStoreExists() throws IOException {
         // Given
-        new S3StateStoreCreator(instanceProperties, dynamoClientV1).create();
+        new S3StateStoreCreator(instanceProperties, dynamoDBClient).create();
         setupTablesPropertiesFile(temporaryFolder, "example-table", S3StateStore.class.getName());
 
         // When
@@ -106,7 +106,7 @@ class NewInstanceValidatorIT extends LocalStackTestBase {
     @Test
     void shouldThrowAnErrorWhenTransactionLogStateStoreExists() throws IOException {
         // Given
-        new DynamoDBStateStoreCreator(instanceProperties, dynamoClientV1).create();
+        new DynamoDBStateStoreCreator(instanceProperties, dynamoDBClient).create();
         setupTablesPropertiesFile(temporaryFolder, "example-table", DynamoDBTransactionLogStateStore.class.getName());
 
         // When
@@ -118,6 +118,6 @@ class NewInstanceValidatorIT extends LocalStackTestBase {
     private void validate() throws IOException {
         Path instancePropertiesPath = temporaryFolder.resolve("instance.properties");
         Files.writeString(instancePropertiesPath, instanceProperties.saveAsString());
-        new NewInstanceValidator(s3Client, dynamoClient).validate(instanceProperties, instancePropertiesPath);
+        new NewInstanceValidator(s3ClientV2, dynamoDbClientV2).validate(instanceProperties, instancePropertiesPath);
     }
 }
